refactor: extract topic query handlers
